flask_server/requirements.txt
Flask
python-dotenv
flask_cors
requests
mypy # type checking
couchdb
pytest